Add a system event emitter
This class publishes system events such as key presses and mouse moves.
Previously this functionality was provided as a function. A class makes
it more manageable and maintainable. In addition, the function required
an EventEmitter instance to be provided as an argument whilst in this
case the class itself is an extension of the EventEmitter


Former-commit-id: abf21e2f421a0d1338fa8665e88e2831e50f2611