Strengthen the condition for adding gui elements to Canvas panel
Previously only the place of placement of the UI element was considered.
This means that UI elements added to the Canvas were allowed to extend outside
the borders of the panel. Now the dimensions of the UI element are considered such
that a UI element added to a Canvas will lie entirely within the Canvas


Former-commit-id: 8c3bd53df8e5e94fd569939834332ae042d03882