Remove border based collision
The entire tile is now either collidable or not. Border based collision
is a bit tricky to detect since we need to know in which direction, the
the tile is being moved and in which direction the colliding entity is
going. We also need to check if the border is collidable on either side
or not. This requires a lot of checking that goes beyond the AABB collision
detection method that the engine uses


Former-commit-id: 5b8ffd8da3ea57c2b1158c632c036ab3db0dc748