Fix adjacency list not reflecting changes to the grid at run-time
The adjacency list always used the tiles initial state and
did not update when the tile was changed from a solid tile
to an empty tile or vice versa.

The same thing happened with obstacles, no update if an obstacle
was added or removed from a tile at run-time. This rendered grid
path finder algorithms inaccurate in situations where a tiles state
changes at run-time


Former-commit-id: 6b4313fffda5ce8d1980ecefd3a734cc80810010