Fix target stopping after collision when there is a valid path to destination
This happens when the target in a TargetGridMover collides with
a solid tile or an obstacle. Tge target completly stops without
trying alternative paths/directions.

The new behavior makes the grid mover try to generate a new path
by in all directions except the direction that results in a
collision. If a path is found the entity is moved in that
direction, otherwise it comes to a complete hault


Former-commit-id: b292a2e5bc767b940966c125c79e4aad61bac56e