Restrict access to public setter fucntion of a Tilemap Tile
Any change to a tilemap tile must be through the tilemap, changing the value
directly may cause inconsistencies in the tilemap. In addition, tilemap tiles
are not meant to be part of the scene. They act as guidance during development
and are usually set to be invisible in release mode. The developer should depend
on sprites and gemotric shapes for visuals


Former-commit-id: 19f2cdfe1edd2abc50e3879968fdb77ebaae7062