Remove Engine parameter from Scene constructor
This parameter is required from the user but the user does not use it.
In addition, the Engine knows when a Scene is added to it so it might
as well inject itself through a member menthod at that point in time.
Therefore the absence of this paramter makes perfect sense and also
makes the user code more cleaner