Remove TODO - Engine throwing std::bad_cast during initialization
The std::bad_any_cast exception was thrown because of a mismatch between the type
of the settings argument and the type the engine was casting to internally. So the
function was working all along. I was the problem. I thought std::any was implicitly
converting const char* arguments to std::string like you would expect in normal situations
but that was not the case, hence the std::bad_any_cast exception