Redefine object ownership
This commit gives the scene manager exclusive ownership of the scene that it
manages - When the scene is popped from the engine it must be destroyed
entirely. In addition, the scene is given exclusive ownership of the objects
that it contains. This solves the problem of having active objects after the
scene is destroyed because at the time of destruction the reference count was
greater than 1