Revert "Fix GridMover::isTargetMoving returning true when movement is frozen"
This reverts commit 6d3b64a9. The "isMoving" and "isMoveFrozen" flags
sound like they are the same thing but actually do two different things.
The "isMoving" flag prevents the targets direction from being changed
while its moving to an adjacent tile whilst the "isMoveFrozen" flag
allows the target to halt its movement en route to the adjacent tile.
When movement is halted, the target must still not change directions
because it has not completed its current adjacent move. So it turns out
"GridMover::isTargetMoving returning true when movement is frozen" was
by design and not an error, hence the rollback.