Fix incorrect ime::Animation frame switch timing when the duration doesn't match the number of frames
When the duration was less than the number of frames, it sometimes resulted in
a frame rate that was not a whole number. The floating point data was truncated
and the animation switched much earlier than it should've. As a result, the
animation finished earlier than expected. When the duration exceeded the number
of frames, the frame rate was always zero which resulted in a frame time of zero
making the animation switch frames every game frame regardless of the duration.