Fix property change callbacks triggering when property value hasn't changed
This commit also fixes the issue of the RAM increasing non-stop when a
RigidBody was instantiated. This is due to the fact that after
every scene update, the scene updates the GameObject that the rigid
body is attached to even if there was no change since the last frame.
Since there was no check for this condition, the GameObject's property
change handlers were always executed, resulting in a huge number of
unnecessary Object copies being made.